pyproject.toml
[project]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = [
    {name = "princisval", email = "christhian.spsilva@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4"  # ✅ Mudança para garantir compatibilidade com flask-jwt-extended
dependencies = [
    "flask",
    "flask-sqlalchemy",
    "flask-migrate",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",  # ✅ Mantemos apenas o flask-jwt-extended
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest",
    "pytest-mock"
] 

[tool.pytest.ini_options]  # ✅ Agora está corretamente separado
minversion = "6.0"
addopts = "-ra -vvv"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
